requirements.txt
selenium==4.8.2
pytest==7.2.2
allure-pytest==2.12.0
webdriver-manager==3.8.5
click==8.1.3

